Add messages - editing of and presentation in unit view. Editing includes ng2-ckeditor. Add som responsive design layout. Simplify routes by removing the children level.

+++ core-bc/composites/svc/src/main/java/se/vgregion/vardplatspusslet/service/BedService.java
@@ -31,7 +31,7 @@ public class BedService {
     @Autowired
     private PatientRepository patientRepository;
 
-    public void deleteBed(String clinicId, String unitId, Integer bedId) {
+    public void deleteBed(String clinicId, String unitId, Long bedId) {
         Clinic clinicRef = clinicRepository.getOne(clinicId);
 
         Unit unit = unitRepository.findUnitByIdIsLikeAndClinicIsLike(unitId, clinicRef);
